TIMOB-10211 Blackberry: Perform KS UI adjustments TIMOB-10433 Blackberry: UI.Label is missing the button showing a background TIMOB-10418 Blackberry: In UI.Button there are no options to change the button text horizontal and vertical by 3positions Reviewers: JP
Changelog:

For the Jira bugs 10433 and 10418 modified code to show the buttons but in
disabled state.
Adjust code for ImageView to work in KS what is supported now.
In several other places modifyed code to not opening the table view at all
if it is not implemented instead of opening and not working.
For example Horizontal Layout and Vertical Layout in Base UI. layout
property is not implemented so clicking on them alert will say that they
are not implemented.

Test:

Run KitchenSink and test the modified components on BB simulator.