TIMOB-10720 Blackberry: Blackberry: UI.WindowsProperties is crashing while pressing "Layout/Mimension Properties" TIMOB-10721 Blackberry: UI.TableView has overlapping buttons in some places TIMOB-10722 Blackberry: UI.AlertDialog has problem with "Basic Alert" button's functionality
Reviewers: JP

Changelog:
For 10720 - removed modifying window code after it was opened.
For 10721  - for some widows for TableView - commented them with alert -
as they are not implemented.
For 10722 - changed Alert Dialog default text to Two buttons
for text field removet workaround for getter.
Uncommented code for Socket to be able to test.

Test:

Run KitchenSink and test the modified components on BB simulator.